+++ packages/plugin-kit/src/index.js
@@ -1,0 +1,12 @@
+/**
+ * @fileoverview Package API
+ * @author Nicholas C. Zakas
+ */
+
+export { ConfigCommentParser } from "./config-comment-parser.js";
+export {
+	CallMethodStep,
+	VisitNodeStep,
+	TextSourceCodeBase,
+	Directive,
+} from "./source-code.js";