+++ packages/core/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [0.6.0](https://github.com/eslint/rewrite/compare/core-v0.5.0...core-v0.6.0) (2024-08-30)
+
+
+### Features
+
+* Add Directive and DirectiveType ([#112](https://github.com/eslint/rewrite/issues/112)) ([e5bbcf1](https://github.com/eslint/rewrite/commit/e5bbcf148874be07d5667f34ed395faaf8c72972))
+
+## [0.5.0](https://github.com/eslint/rewrite/compare/core-v0.4.0...core-v0.5.0) (2024-08-27)
+
+
+### Features
+
+* Add plugin-kit package ([#99](https://github.com/eslint/rewrite/issues/99)) ([09ce430](https://github.com/eslint/rewrite/commit/09ce43073760b69a3bcca89f99793549cd566bf6))
+
 ## [0.4.0](https://github.com/eslint/rewrite/compare/core-v0.3.0...core-v0.4.0) (2024-08-12)